✨ Updated NodeInfo

+++ Fedodo.BE.ActivityPub/Model/NodeInfo/Usage.cs
@@ -3,6 +3,6 @@ namespace Fedodo.BE.ActivityPub.Model.NodeInfo;
 public class Usage
 {
     public Users? Users { get; set; }
-    public int LocalPosts { get; set; }
-    public int LocalComments { get; set; }
+    public long LocalPosts { get; set; }
+    public long LocalComments { get; set; }
 }
