+++ Fedodo.BE.ActivityPub/Model/NodeInfo/Users.cs
@@ -2,7 +2,7 @@ namespace Fedodo.BE.ActivityPub.Model.NodeInfo;
 
 public class Users
 {
-    public int Total { get; set; }
+    public long Total { get; set; }
     public int ActiveMonth { get; set; }
     public int ActiveHalfyear { get; set; }
 }
